requirements.txt
pympq>=0.1.1
pillow>=8.0.0